+++ jobs/publish.sh
@@ -23,6 +23,7 @@ docker build -t skelebot/python-base:3.7 base-images/python-base/3.7/
 docker build -t skelebot/python-base:3.8 base-images/python-base/3.8/
 docker build -t skelebot/python-base:3.9 base-images/python-base/3.9/
 docker build -t skelebot/python-base:3.10 base-images/python-base/3.10/
+docker build -t skelebot/python-base:3.11 base-images/python-base/3.11/
 # Newer Docker clients will push latest by default and will need to do `docker push skelebot/python-base --all-tags`
 docker push skelebot/python-base